requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-jazzmin==2.6.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
